requirements.txt
# Stock Analysis & Prediction Platform Dependencies

# Data Collection & Processing
yfinance==0.2.28
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
APScheduler==3.10.4

# Web Framework
Flask==2.2.5  # Compatible with Dash
Flask-SQLAlchemy==3.1.1
flask-cors==4.0.0

# Visualization
plotly==5.18.0
dash==2.14.0

# Analysis
scikit-learn==1.3.0
pandas-ta==0.3.14b0

# Performance Optimization
redis==5.0.1
celery==5.3.4

# Machine Learning (Optional)
# tensorflow==2.15.0
# keras==2.15.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1 